[ISSUE #8238]♻️Extract Proxy Core ports and ingress
