+++ rocketmq-proxy-core/Cargo.toml
@@ -32,8 +32,10 @@ default = []
 rocketmq-error = { workspace = true }
 rocketmq-model = { workspace = true }
 rocketmq-protocol = { workspace = true }
+rocketmq-runtime = { workspace = true }
 rocketmq-transport = { workspace = true, default-features = false }
 
+async-trait = "0.1"
 dashmap.workspace = true
 serde.workspace = true
 thiserror.workspace = true
