+++ rocketmq-proxy-core/src/grpc/service/producer.rs
@@ -12,4 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub(crate) use rocketmq_proxy_core::grpc::middleware::ingress_context_interceptor;
+use crate::ProxyMessage;
+
+const TRANSACTION_PREPARED_PROPERTY: &str = "TRAN_MSG";
+
+pub fn is_transaction_message(message: &ProxyMessage) -> bool {
+    message.property(TRANSACTION_PREPARED_PROPERTY).is_some()
+}
