+++ rocketmq-proxy/src/lib.rs
@@ -29,6 +29,7 @@ pub mod context;
 pub mod error;
 pub mod grpc;
 pub mod local;
+mod message;
 pub mod observability;
 pub mod processor;
 pub mod proto;
@@ -98,6 +99,8 @@ pub use processor::TransactionSource;
 pub use processor::UpdateOffsetPlan;
 pub use processor::UpdateOffsetRequest;
 pub use remoting::ProxyRemotingDispatcher;
+pub use rocketmq_proxy_core::ProxyMessage;
+pub use rocketmq_proxy_core::ProxyMessageExt;
 pub use service::AssignmentService;
 pub use service::ClusterServiceManager;
 pub use service::ConsumerService;
